# src/main/resources/openapi/clientes-api.yml
openapi: 3.0.3
info:
  title: API Clientes
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/clientes:
    get:
      summary: Filtrar clientes
      operationId: filtrarCliente
      parameters:
        - name: nombre
          in: query
          required: false
          schema:
            type: string
        - name: nroDocumento
          in: query
          required: false
          schema:
            type: string
        - name: idTipo
          in: query
          required: false
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de clientes paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClienteResponseDTO'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  number:
                    type: integer
                  size:
                    type: integer

    post:
      summary: Crear cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteRequestDTO'
      responses:
        '200':
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponseDTO'
...
components:
  schemas:
    ClienteRequestDTO:
      type: object
      properties:
        nombre:
          type: string
        nroDocumento:
          type: string
        email:
          type: string
        celular:
          type: string
        tipoCliente:
          type: integer
      required:
        - nombre
        - nroDocumento
        - email
        - celular
        - tipoCliente
    ClienteResponseDTO:
      type: object
      properties:
        idCliente:
          type: integer
        nombre:
          type: string
        nroDocumento:
          type: string
        email:
          type: string
        celular:
          type: string
        tipoCliente:
          type: integer
...